Replacing Variant with IndexArray as arguments to BasicArray class -- makes the interface to BasicArray independent of the Variant class.  Also switching to doubles as indexing types is more consistent with the new type system and avoids an extra typecast for each operation.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/array-rework@3171 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/demo/Global.hpp
@@ -3,7 +3,7 @@
 
 #include <QtGlobal>
 
-typedef qint32 index_t;
+typedef double index_t;
 const int NDims = 6;
 typedef QString Exception;
